Rubocop実行によるコード修正

+++ app/controllers/items_controller.rb
@@ -23,10 +23,9 @@ class ItemsController < ApplicationController
   end
 
   def edit
-    if Order.exists?(item_id: @item.id)
-      redirect_to action: :index
-    end
+    redirect_to action: :index if Order.exists?(item_id: @item.id)
     return if current_user.id == @item.user_id
+
     redirect_to action: :index
   end
 
@@ -40,10 +39,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if current_user.id == item.user_id
-      item.destroy
-    end
-      redirect_to root_path
+    item.destroy if current_user.id == item.user_id
+    redirect_to root_path
   end
 
   private
